+++ examples/storybook/stories/ContentAPI/Panels/index.ts
@@ -1,0 +1,1 @@
+export * from './Paragraph';
