+++ examples/storybook/stories/ContentAPI/types.ts
@@ -1,8 +1,13 @@
-interface EditPanelProps<T> {
-  onAdd: (args: T) => void;
+import { setupContentBuilder } from 'ricos-content/libs/Content';
+
+export type BuilderFunctions = Omit<ReturnType<typeof setupContentBuilder>, 'RicosContentBuilder'>;
+export type AddFunctor = <T extends keyof BuilderFunctions>(
+  element: T,
+  args: Omit<Parameters<BuilderFunctions[T]>['0'], 'content'>
+) => void;
+
+export interface EditPanelProps<T extends keyof BuilderFunctions> {
+  addFunc: (args: Omit<Parameters<BuilderFunctions[T]>['0'], 'content'>) => void;
 }
 
-type EditTextProps = EditPanelProps<{
-  text?: string;
-  alignment?: string;
-}>;
+export interface ParagraphPanel extends EditPanelProps<'addParagraph'> {}
